requirements.txt
pandas>=1.5.0
numpy>=1.23.0
requests>=2.28.0
pyyaml>=6.0
ta>=0.10.0
python-dotenv>=0.21.0